Implemented Questionary
NOTE: a simplified IDE terminal will give an error if you try to use questionary in it.
PyCharm allows you to a) go into the run configuration and check the "emulate terminal in output console" option, or b) open the terminal tab in the bottom bar and type the command to run it manually.
Either one of those will allow you to run the application like normal.