Improved accuracy
The main things that would be left are pair splitting and insurance or even money, which are often not even options at certain casinos. Oh, and another major things I've forgotten or missed. Also, I'm sure there's still bugs or minor things that can be done to improve the experience