Improved early stopping message
I was unhappy with the fact that the mention of being able to start the game early stayed at the top of the screen, potentially being hidden by being among several other messages, so ANSI escape sequences were used to make sure the message stays at the bottom until the game starts.

There was also an issue with the input function never truly ending, and the server host therefore being able to type at any point during the session. This was fixed by implementing the inputimeout library and using yet more ANSI escape sequences.

It is still not perfect, but anything more would require significantly more time invested, and the deadline is approaching.