Removed player limit caused by code lmitations
Originally, we had no way of handling late connections if there were more than 10 players, as we had no way of excluding the digits at the end of a player's name that prevent name repetition if we they were not fixed length. Now, the player's original name without the digits is saved as well, to make getting that name easy.